requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
black==24.8.0
blinker==1.8.2
click==8.1.7
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fastapi==0.115.2
Flask==3.0.3
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==3.0.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pdfkit==1.0.0
platformdirs==4.2.2
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.39.2
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.31.1
Werkzeug==3.0.4